level_statistics: Prevent potential endless loop in init_level_base_voronoi()
init_level_base_voronoi() tries to put a fixed number of points on the level
and positioning them at a certain distance from each other. This could result
in an endless loop as only the current point was repositioned.

The code now will give up and continue with a suboptimal point as later on the
level will be checked for certain properties and if failing that, thrown away
completely anyway.